Make the project compile again
