+++ Rocket.Chat/Managers/Model/SubscriptionManager/SubscriptionManager.swift
@@ -66,7 +66,7 @@ struct SubscriptionManager {
                     list?.forEach { object in
                         if let rid = object["_id"].string {
                             if let subscription = Subscription.find(rid: rid, realm: realm) {
-                                subscription.mapRoom(object)
+                                subscription.mapRoom(object, realm: realm)
                                 subscriptions.append(subscription)
                             }
                         }
@@ -75,7 +75,7 @@ struct SubscriptionManager {
                     updated?.forEach { object in
                         if let rid = object["_id"].string {
                             if let subscription = Subscription.find(rid: rid, realm: realm) {
-                                subscription.mapRoom(object)
+                                subscription.mapRoom(object, realm: realm)
                                 subscriptions.append(subscription)
                             }
                         }
@@ -188,8 +188,7 @@ struct SubscriptionManager {
             currentRealm?.execute({ (realm) in
                 if let rid = object["_id"].string {
                     if let subscription = Subscription.find(rid: rid, realm: realm) {
-                        subscription.mapRoom(object)
-
+                        subscription.mapRoom(object, realm: realm)
                         realm.add(subscription, update: true)
                     }
                 }
