+++ Rocket.Chat/Models/Subscription/Subscription.swift
@@ -57,42 +57,6 @@ final class Subscription: BaseModel {
     let messages = LinkingObjects(fromType: Message.self, property: "subscription")
 }
 
-extension Subscription {
-
-    func lastMessageText() -> String {
-        guard
-            let lastMessage = roomLastMessage,
-            let userLastMessage = lastMessage.user
-        else {
-            return "No message"
-        }
-
-        var text = lastMessage.text
-
-        let isFromCurrentUser = userLastMessage.identifier == AuthManager.currentUser()?.identifier
-        let isOnlyAttachment = text.isEmpty && lastMessage.attachments.count > 0
-
-        if isOnlyAttachment {
-            text = " sent an attachment"
-        } else {
-            if !isFromCurrentUser {
-                text = ": \(text)"
-            }
-        }
-
-        if isFromCurrentUser && isOnlyAttachment {
-            text = "You\(text)"
-        }
-
-        if !isFromCurrentUser {
-            text = "\(userLastMessage.displayName())\(text)"
-        }
-
-        return text
-    }
-
-}
-
 // MARK: Failed Messages
 
 extension Subscription {
